requirements.txt
scipy==1.16.2
tensorflow-cpu>=2.12,<3